+++ products/jtree.node.js
@@ -2105,6 +2105,12 @@ class TreeNode extends AbstractNode {
     this.getChildren().forEach(fn)
     return this
   }
+  // Recurse if predicate passes
+  deepVisit(predicate) {
+    this.forEach(node => {
+      if (predicate(node) !== false) node.deepVisit(predicate)
+    })
+  }
   // todo: protected?
   _clearIndex() {
     delete this._index
@@ -2992,7 +2998,7 @@ TreeNode.iris = `sepal_length,sepal_width,petal_length,petal_width,species
 4.9,2.5,4.5,1.7,virginica
 5.1,3.5,1.4,0.2,setosa
 5,3.4,1.5,0.2,setosa`
-TreeNode.getVersion = () => "45.0.1"
+TreeNode.getVersion = () => "45.1.0"
 class AbstractExtendibleTreeNode extends TreeNode {
   _getFromExtended(firstWordPath) {
     const hit = this._getNodeFromExtended(firstWordPath)