+++ products/stump.nodejs.js
@@ -188,6 +188,9 @@ htmlTagNode
    const oneLinerWords = this.getWordsFrom(1)
    return oneLinerWords.length ? oneLinerWords.join(" ") : ""
   }
+  getTextContent() {
+    return this._getOneLiner()
+  }
   shouldCollapse() {
    return this.has("stumpCollapse")
   }
@@ -338,6 +341,7 @@ htmlAttributeNode
   _toHtml() {
    return ""
   }
+  getTextContent() {return ""}
   getAttribute() {
    return \` \${this.getFirstWord()}="\${this.getContent()}"\`
   }
@@ -357,6 +361,8 @@ lineOfHtmlContentNode
  boolean isTileAttribute true
  catchAllNodeType lineOfHtmlContentNode
  catchAllCellType anyHtmlContentCell
+ javascript
+  getTextContent() {return this.getLine()}
 bernNode
  boolean isTileAttribute true
  todo Rename this node type
@@ -366,6 +372,7 @@ bernNode
   _toHtml() {
    return this.childrenToString()
   }
+  getTextContent() {return ""}
  cells bernKeywordCell`)
       return this._cachedGrammarProgramRoot
     }
@@ -700,6 +707,9 @@ bernNode
       const oneLinerWords = this.getWordsFrom(1)
       return oneLinerWords.length ? oneLinerWords.join(" ") : ""
     }
+    getTextContent() {
+      return this._getOneLiner()
+    }
     shouldCollapse() {
       return this.has("stumpCollapse")
     }
@@ -870,6 +880,9 @@ bernNode
     _toHtml() {
       return ""
     }
+    getTextContent() {
+      return ""
+    }
     getAttribute() {
       return ` ${this.getFirstWord()}="${this.getContent()}"`
     }
@@ -891,6 +904,9 @@ bernNode
     get isTileAttribute() {
       return true
     }
+    getTextContent() {
+      return this.getLine()
+    }
   }
 
   class bernNode extends jtree.GrammarBackedNode {
@@ -905,6 +921,9 @@ bernNode
     }
     _toHtml() {
       return this.childrenToString()
+    }
+    getTextContent() {
+      return ""
     }
   }