+++ treeComponentFramework/sweepercraft/SweeperCraft.ts
@@ -726,8 +726,7 @@ githubTriangleComponent`
       })
   }
 
-  private _loadFromHash(stumpNode: any) {
-    const link = location.hash.replace(/^\#/, "")
+  private _loadFromHash(stumpNode: any, link: string) {
     let board
     if (!link) board = SweeperCraftGame.getRandomBoard(9, 9, 10)
     else board = SweeperCraftGame.boardFromPermalink(link)
@@ -755,33 +754,39 @@ githubTriangleComponent`
     return classes
   }
 
-  private _firstRender(stumpNode: any) {
+  private _setupBrowser(stumpNode: any) {
     window.addEventListener("error", err => {
       jQuery("#errors").html(`Something went wrong: ${err.message}. <a href=''>Refresh</a>`)
     })
 
+    const willowProgram = this.getWillowProgram()
+
     const keyboardShortcuts: any = this._getKeyboardShortcuts()
     Object.keys(keyboardShortcuts).forEach(key => {
-      this.getWillowProgram()
-        .getMousetrap()
-        .bind(key, function(evt: any) {
-          keyboardShortcuts[key]()
-          // todo: handle the below when we need to
-          if (evt.preventDefault) evt.preventDefault()
-          return false
-        })
+      willowProgram.getMousetrap().bind(key, function(evt: any) {
+        keyboardShortcuts[key]()
+        // todo: handle the below when we need to
+        if (evt.preventDefault) evt.preventDefault()
+        return false
+      })
     })
 
     Figlet.loadFont("banner", FontsBanner)
 
     window.addEventListener("hashchange", () => {
       console.log("hashchange")
-      this._loadFromHash(stumpNode)
+      this._loadFromHash(stumpNode, willowProgram.getHash().replace(/^\#/, ""))
     })
+  }
+
+  private _firstRender(stumpNode: any) {
+    if (!this.isNodeJs()) this._setupBrowser(stumpNode)
+
+    const willowProgram = this.getWillowProgram()
 
     // Initialize first game
-    if (SweeperCraftGame.isValidPermalink(location.hash.replace(/^#/, ""))) this._loadFromHash(stumpNode)
-    else location.hash = SweeperCraftGame.toPermalink(SweeperCraftGame.getRandomBoard(9, 9, 10))
+    if (SweeperCraftGame.isValidPermalink(willowProgram.getHash().replace(/^#/, ""))) this._loadFromHash(stumpNode, willowProgram.getHash().replace(/^#/, ""))
+    else willowProgram.setHash(SweeperCraftGame.toPermalink(SweeperCraftGame.getRandomBoard(9, 9, 10)))
   }
 }
 
@@ -792,7 +797,7 @@ abstract class AbstractSweeperCraftComponent extends AbstractTreeComponent {
 class headerComponent extends AbstractSweeperCraftComponent {
   treeComponentDidMount() {
     super.treeComponentDidMount()
-    this._initTimerInterval()
+    if (!this.isNodeJs()) this._initTimerInterval()
   }
 
   treeComponentWillUnmount() {